Round the shapes in the tessellator instead
