+++ crates/egui/src/introspection.rs
@@ -144,6 +144,8 @@ impl Widget for &mut epaint::TessellationOptions {
                 coarse_tessellation_culling,
                 prerasterized_discs,
                 round_text_to_pixels,
+                round_line_segments_to_pixels,
+                round_rects_to_pixels,
                 debug_paint_clip_rects,
                 debug_paint_text_rects,
                 debug_ignore_clip_rects,
@@ -179,13 +181,22 @@ impl Widget for &mut epaint::TessellationOptions {
 
             ui.checkbox(validate_meshes, "Validate meshes").on_hover_text("Check that incoming meshes are valid, i.e. that all indices are in range, etc.");
 
+            ui.collapsing("Align to pixel grid", |ui| {
+                ui.checkbox(round_text_to_pixels, "Text")
+                    .on_hover_text("Most text already is, so don't expect to see a large change.");
+
+                ui.checkbox(round_line_segments_to_pixels, "Line segments")
+                    .on_hover_text("Makes line segments appear crisp on any display.");
+
+                ui.checkbox(round_rects_to_pixels, "Rectangles")
+                    .on_hover_text("Makes line segments appear crisp on any display.");
+            });
+
             ui.collapsing("Debug", |ui| {
                 ui.checkbox(
                     coarse_tessellation_culling,
                     "Do coarse culling in the tessellator",
                 );
-                ui.checkbox(round_text_to_pixels, "Align text positions to pixel grid")
-                    .on_hover_text("Most text already is, so don't expect to see a large change.");
 
                 ui.checkbox(debug_ignore_clip_rects, "Ignore clip rectangles");
                 ui.checkbox(debug_paint_clip_rects, "Paint clip rectangles");
