+++ crates/egui_demo_lib/tests/snapshots/rendering_test/dpi_1.25.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:04787907e9018d97daa5ea653f8907314746b4d2a83b8534c78ffdeaf207ca59
-size 742732
+oid sha256:c2d0bae4ac670dbfcd709b6e0f0155a299812b8c9cdb40379af175a2d022cfa8
+size 742727
