+++ crates/egui_demo_lib/tests/snapshots/rendering_test/dpi_1.50.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:bfa1710dde28dcac50c75358aae1d89db182f7189f7365f962a54063c2ce1dac
-size 884450
+oid sha256:968f64f2d7db93a632aaa44818766af0d1150653377fdbbe545c7c530aa653db
+size 884441
