+++ crates/egui_demo_lib/tests/snapshots/rendering_test/dpi_1.67.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:2e6ab27392c9800010b4a7c04b3ae7b85839953388ed3f71bfa297f454b692cf
-size 993108
+oid sha256:147ce8ec08bcf6e4c06a46eaaf4a4cef8c268876725cd96648351f0f51610a6d
+size 993098
