+++ crates/egui_demo_lib/tests/snapshots/rendering_test/dpi_1.75.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:d757545fe3b16edc23a698361952cbc5deaf9651aee391c3cc595732df2c8e28
-size 1085198
+oid sha256:e9c5b829d3462c9a2fdf29c4e7d9e1e62fbfa7656c104e8f481364b659957e3a
+size 1085193
